requirements.txt
ffmpeg-python==0.2.0
future==0.18.2
lxml==4.6.2
Pillow==8.0.1
python-pptx==0.6.18
XlsxWriter==1.3.7
youtube-dl==2020.11.29
